requirements.txt
appdirs==1.4.3
colorthief==0.2.1
Django==1.10.6
django-appconf==1.0.2
django-cors-headers==2.0.2
django-imagekit==4.0
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
